feat:新增Enum代码生成器

+++ Level4/Materal.BaseCore/Src/Materal.BaseCore.PresentationModel/IQueryRequestModel.cs
@@ -8,10 +8,10 @@
         /// <summary>
         /// 每页数量
         /// </summary>
-        public int PageSize { get; set; }
+        public long PageSize { get; set; }
         /// <summary>
         /// 页码
         /// </summary>
-        public int PageIndex { get; set; }
+        public long PageIndex { get; set; }
     }
 }
